Remove query subtyping trickery
Instead of attaching query result types to individuals, the
responsibility is now moved to the specified types. You can just check
that all elements of a set of individuals match a type instead of having
to deal with individuals that magically have two different types
depending on how you look at them.