Add caching to memory and disk
Everything the knowledge base does gets cached in memory and written to
disk when the process ends. Usually this leads to the knowledge base not
even being instantiated on subsequent runs.

The data source file is flocked so that multiple processes don't step on
each other's toes reading and writing the cache file. The test suite can
still run in parallel that way.

The location for the cache file is still kinda weird though, it's just
stashed in the current directory. The JSON::Fast requirement is also
slightly annoying, since it'd be the only non-core dependency.